Add files via upload
The thing that is cool about Google Dorks is that it can be an amazing tool for searching stuff. In this example, the Google Dork query searches for web pages on the example.com domain that have "index.of" in the title. The search() function returns up to 10 results, which are then printed to the console. 

***Note that  the 'query' variable can be modified to for other types of information, such as specific file types, directories, or sensitive information like passwords or login credentials. This is according to my research, not that I know how to exactly do that.