Create Search_Query
Ok, so this is a really cool one. This is the code itself to Google Dorks, and you would paste DIRECTLY in the search bar of Google. This Dork searches for files with the extension .log that contain the word "access" in the URL and any of the following keywords in the file content: "password", "user", "username", "email", "login", or "pass". This is used for security purposes by experts. This is incredible!